requirements.txt
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
html2markdown==0.1.7
idna==2.10
jdcal==1.4.1
jieba==0.42.1
lxml==4.5.2
Naked==0.1.31
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.2
pycryptodome==3.9.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
tomd==0.1.3
urllib3==1.25.10
user-agent==0.1.9
xlrd==1.2.0
xlwt==1.3.0
